Rename fluent filters to remove Soql reference
