+++ src/fluent-query/filters/and-where-filter.ts
@@ -1,10 +1,8 @@
-import { Operator, WhereFilter, WhereOperator } from '../../soql-query-builder';
-import { Comparitor } from "../../soql-query-builder/clauses/where/comparitor";
-import { WhereValue } from "../../soql-query-builder/clauses/where/where-value";
+import { Comparitor, Operator, WhereFilter, WhereOperator, WhereValue } from '../../soql-query-builder';
 import { IFilteredQueryable } from '../ifilteredqueryable';
-import { SoqlWhereFilter } from './soql-where-filter';
+import { BasicWhereFilter } from './basic-where-filter';
 
-export class SoqlOrWhereFilter<TEntity, TValue> extends SoqlWhereFilter<TEntity, TValue> {
+export class AndWhereFilter<TEntity, TValue> extends BasicWhereFilter<TEntity, TValue> {
 
   public equals(value: TValue): IFilteredQueryable<TEntity> {
     if (value === null) {
@@ -12,7 +10,7 @@ export class SoqlOrWhereFilter<TEntity, TValue> extends SoqlWhereFilter<TEntity,
     }
 
     const filter = new WhereFilter(this.column, Comparitor.Equals, new WhereValue(value));
-    return this.query.addFilter(new WhereOperator(Operator.Or), filter);
+    return this.query.addFilter(new WhereOperator(Operator.And), filter);
   }
 
   public greaterThan(value: TValue): IFilteredQueryable<TEntity> {
@@ -21,7 +19,7 @@ export class SoqlOrWhereFilter<TEntity, TValue> extends SoqlWhereFilter<TEntity,
     }
 
     const filter = new WhereFilter(this.column, Comparitor.GreaterThan, new WhereValue(value));
-    return this.query.addFilter(new WhereOperator(Operator.Or), filter);
+    return this.query.addFilter(new WhereOperator(Operator.And), filter);
   }
 
   public lessThan(value: TValue): IFilteredQueryable<TEntity> {
@@ -30,12 +28,12 @@ export class SoqlOrWhereFilter<TEntity, TValue> extends SoqlWhereFilter<TEntity,
     }
 
     const filter = new WhereFilter(this.column, Comparitor.LessThan, new WhereValue(value));
-    return this.query.addFilter(new WhereOperator(Operator.Or), filter);
+    return this.query.addFilter(new WhereOperator(Operator.And), filter);
   }
 
   public isNull(): IFilteredQueryable<TEntity> {
     const filter = new WhereFilter(this.column, Comparitor.IsNull);
-    return this.query.addFilter(new WhereOperator(Operator.Or), filter);
+    return this.query.addFilter(new WhereOperator(Operator.And), filter);
   }
 
   public isNotNull(): IFilteredQueryable<TEntity> {
