+++ src/fluent-query/filters/or-geometry-filter.ts
@@ -11,7 +11,7 @@ import { IFilteredQueryable } from '../ifilteredqueryable';
 import { IInternalQuery } from '../iinternalquery';
 import { IGeometryFilter } from './igeometryfilter';
 
-export class SoqlOrGeometryFilter<TEntity> implements IGeometryFilter<TEntity> {
+export class OrGeometryFilter<TEntity> implements IGeometryFilter<TEntity> {
 
   public constructor(protected readonly query: IInternalQuery<TEntity>, protected readonly column: Column) {
     if (!query) {
