+++ src/fluent-query/filters/or-location-filter.ts
@@ -9,7 +9,7 @@ import { IFilteredQueryable } from '../ifilteredqueryable';
 import { IInternalQuery } from '../iinternalquery';
 import { ILocationFilter } from './ilocationfilter';
 
-export class SoqlOrLocationFilter<TEntity> implements ILocationFilter<TEntity> {
+export class OrLocationFilter<TEntity> implements ILocationFilter<TEntity> {
 
   public constructor(protected readonly query: IInternalQuery<TEntity>, protected readonly column: Column) {
     if (!query) {
