+++ src/fluent-query/filters/or-where-filter.ts
@@ -1,8 +1,10 @@
-import { Comparitor, Operator, WhereFilter, WhereOperator, WhereValue } from '../../soql-query-builder';
+import { Operator, WhereFilter, WhereOperator } from '../../soql-query-builder';
+import { Comparitor } from "../../soql-query-builder/clauses/where/comparitor";
+import { WhereValue } from "../../soql-query-builder/clauses/where/where-value";
 import { IFilteredQueryable } from '../ifilteredqueryable';
-import { SoqlWhereFilter } from './soql-where-filter';
+import { BasicWhereFilter } from './basic-where-filter';
 
-export class SoqlAndWhereFilter<TEntity, TValue> extends SoqlWhereFilter<TEntity, TValue> {
+export class OrWhereFilter<TEntity, TValue> extends BasicWhereFilter<TEntity, TValue> {
 
   public equals(value: TValue): IFilteredQueryable<TEntity> {
     if (value === null) {
@@ -10,7 +12,7 @@ export class SoqlAndWhereFilter<TEntity, TValue> extends SoqlWhereFilter<TEntity
     }
 
     const filter = new WhereFilter(this.column, Comparitor.Equals, new WhereValue(value));
-    return this.query.addFilter(new WhereOperator(Operator.And), filter);
+    return this.query.addFilter(new WhereOperator(Operator.Or), filter);
   }
 
   public greaterThan(value: TValue): IFilteredQueryable<TEntity> {
@@ -19,7 +21,7 @@ export class SoqlAndWhereFilter<TEntity, TValue> extends SoqlWhereFilter<TEntity
     }
 
     const filter = new WhereFilter(this.column, Comparitor.GreaterThan, new WhereValue(value));
-    return this.query.addFilter(new WhereOperator(Operator.And), filter);
+    return this.query.addFilter(new WhereOperator(Operator.Or), filter);
   }
 
   public lessThan(value: TValue): IFilteredQueryable<TEntity> {
@@ -28,12 +30,12 @@ export class SoqlAndWhereFilter<TEntity, TValue> extends SoqlWhereFilter<TEntity
     }
 
     const filter = new WhereFilter(this.column, Comparitor.LessThan, new WhereValue(value));
-    return this.query.addFilter(new WhereOperator(Operator.And), filter);
+    return this.query.addFilter(new WhereOperator(Operator.Or), filter);
   }
 
   public isNull(): IFilteredQueryable<TEntity> {
     const filter = new WhereFilter(this.column, Comparitor.IsNull);
-    return this.query.addFilter(new WhereOperator(Operator.And), filter);
+    return this.query.addFilter(new WhereOperator(Operator.Or), filter);
   }
 
   public isNotNull(): IFilteredQueryable<TEntity> {
